Event Detail View UI
Essentially copied the picture Justin drew for the event detail view screen. The notification button and back button works. The sign up button works (no functionality). Those are the only buttons so far. We may want to be able to click on all the profile pictures to open that user's profile.