Rounded edges, horizontal layouts, and shadows (Oh my!)
Gave most things a rounded edge. Moved the sign up button to be on the bottom bar where you can send messages. Changed the participants list to be horizontal instead of vertical. You can also click on any participant to open their profile, the event creator image is clickable, and the map image is clickable (no functionality). I Put shadows on things becaused it looked kinda neat. Separted some of the code into separate functions. Currently in the same page, but most likely will be moved so that they can be used anywhere.