Getting User Data + Trying To Store Lists in Firebase
Added stuff to get all the registered users. The event detail example auto fills the participants with all registered users. Also added functionality to the participant buttons. It will show the profile of whatever user was clicked on. I made the classes and functions to be able to create an event. I haven't used it yet because I have been having trouble trying to store lists to firebase.